+++ NEWS.md
@@ -1,7 +1,9 @@
 # misha (development version)
 
+* Added `gtrack.create_dense` function - creates a dense track from an intervals and values.
 * Fixed compilation error in old versions of macOS (`clock_gettime` is missing).
 * Fixed a bug in `gtrack.import_bigwig`: `intern` argument was not passed to `system` calls.
+
 
 # misha 4.3.6